+++ test/resources/unnecessary_nullable_analyzer/unnecessary_nullable_example.dart
@@ -13,7 +13,8 @@ void main() {
 
   withMethods
     ..someMethod(null)
-    ..someMethod('value');
+    ..someMethod('value')
+    ..tearOff('str');
 
   withMethods.alwaysNonNullable('anotherValue');
   withMethods.ignoredAlwaysNonNullable('anotherValue');
@@ -44,6 +45,8 @@ void main() {
   multipleParametersWithOptional('name', 1, 'secondName');
 
   MyWidget(GlobalKey());
+
+  AnotherWidget(onSubmit: withMethods.tearOff);
 }
 
 class _Test {